backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.1
pydantic==2.6.4